pyproject.toml
[project]
name = "reddit-image-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "certifi==2024.8.30",
    "charset-normalizer==3.3.2",
    "google-api-python-client>=2.169.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "greenlet==3.1.1",
    "idna==3.10",
    "jmespath==1.0.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "s3transfer==0.10.2",
    "six==1.16.0",
    "sqlalchemy==2.0.35",
    "typing-extensions==4.12.2",
    "urllib3==2.2.3",
]
